requirements.txt
# Core Python utilities
python-dotenv==1.0.1

# Flask web framework
Flask==3.0.2

# LangChain ecosystem
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-pinecone==0.1.0

# PDF loading
PyPDF2==3.0.1

# Embeddings
sentence-transformers==3.2.1

# Vector database
pinecone-client==4.1.0

# YAML parsing
PyYAML==6.0.2

# Groq client
groq==0.9.0

# Testing
pytest==8.1.1